Moving grid.svelte from pages to components
